// src/app/interfaces/interfaces.ts
// Generated by https://quicktype.io

export interface RespuestaPost {
    ok: boolean;
    pagina: number;
    post: Post[];
}

export interface Post {
    _id?: string;
    mensaje?: string;
    img?: string[];
    coords?: string;
    usuario?: Usuario;
    created?: string;
}

export interface Usuario {
    _id?: string;
    nombre?: string;
    avatar?: string;
    email?: string;
    password?: string;
}